+++ pom.xml
@@ -6,7 +6,7 @@
 
     <groupId>com.kttdevelopment</groupId>
     <artifactId>simplehttpserver</artifactId>
-    <version>03.05.07</version>
+    <version>03.05.08</version>
     <packaging>jar</packaging>
 
     <url>https://github.com/Ktt-Development/simplehttpserver</url>
