+++ src/main/java/com/kttdevelopment/simplehttpserver/var/RequestMethod.java
@@ -7,7 +7,6 @@ package com.kttdevelopment.simplehttpserver.var;
  * @version 01.00.00
  * @author Ktt Development
  */
-@SuppressWarnings("SpellCheckingInspection")
 public enum RequestMethod {
     GET,
     HEAD,
