+++ src/test/java/handlers/temporary/TemporaryTimeTest.java
@@ -1,8 +1,8 @@
-package handlers;
+package handlers.temporary;
 
 import com.kttdevelopment.simplehttpserver.*;
-import com.kttdevelopment.simplehttpserver.handler.RedirectHandler;
-import com.kttdevelopment.simplehttpserver.var.HttpCode;
+import com.kttdevelopment.simplehttpserver.handler.TemporaryHandler;
+import com.sun.net.httpserver.HttpExchange;
 import org.junit.Assert;
 import org.junit.Test;
 
@@ -11,19 +11,21 @@ import java.net.URI;
 import java.net.http.*;
 import java.util.concurrent.ExecutionException;
 
-public final class RedirectHandlerTests {
+public final class TemporaryTimeTest {
 
     @Test
-    public final void redirectToGoogle() throws IOException, ExecutionException, InterruptedException{
+    public final void testTime() throws IOException, InterruptedException, ExecutionException{
         final int port = 8080;
 
         final SimpleHttpServer server = SimpleHttpServer.create(port);
 
         final String context = "";
-        server.createContext(context,new RedirectHandler("https://www.google.com/"));
+        server.createContext(context,new TemporaryHandler(server, HttpExchange::close, 1000));
         server.start();
 
         Assert.assertFalse("Server did not contain a temporary context", server.getContexts().isEmpty());
+
+        Thread.sleep(2000);
 
         final String url = "http://localhost:" + port + context;
 
@@ -31,12 +33,10 @@ public final class RedirectHandlerTests {
             .uri(URI.create(url))
             .build();
 
-        final HttpClient client = HttpClient.newBuilder().followRedirects(HttpClient.Redirect.ALWAYS).build();
-
-        final int response = client.sendAsync(request, HttpResponse.BodyHandlers.ofString())
+        HttpClient.newHttpClient().sendAsync(request, HttpResponse.BodyHandlers.ofString())
             .thenApply(HttpResponse::statusCode).get();
 
-        Assert.assertEquals("Client responded with redirect code (302 HTTP FOUND) not 200 HTTP OK",HttpCode.HTTP_OK, response);
+        Assert.assertTrue("Server did not remove temporary context", server.getContexts().isEmpty());
 
         server.stop();
     }
