+++ src/test/java/simplehttpserver/SimpleHttpServerContextTests.java
@@ -3,7 +3,6 @@ package simplehttpserver;
 import com.kttdevelopment.simplehttpserver.*;
 import com.kttdevelopment.simplehttpserver.handler.RootHandler;
 import com.sun.net.httpserver.HttpExchange;
-import com.sun.net.httpserver.HttpServer;
 import org.junit.Assert;
 import org.junit.Test;
